requirements.txt
fastapi==0.109.2
numpy==1.26.4
tensorflow==2.15.0
tensorflow-hub==0.16.1
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==10.0.0
azure-ai-vision-imageanalysis==1.0.0b2